refactor expanded flow graph layout modules
